Move type relations into submodule in rustc_infer
